Renamed unsubscriber class
